Fix inputs and outputs to be valid JSON again
JSON-RPC 2.0 requires inputs and outputs to be plain JSON, but flatted
uses a different serialization algorithm, even for non-circular objects,
and this breaks the compatibility. Before flatted, circular-json was
used and it seemed to be compatible with JSON for non-circular objects.
So it seems like reverting to it could be the right option here but it's
not guaranteed to be JSON compatible either. So, for the short term, it
looks like it's the best option to remove circular object support. A
long term solution without breaking JSON-RPC compatibility could be
developed after.

Before this change, only 26 of 75 tests were passing, now all of them
pass.